simplefs: subscribe to empty TLFs after they are created
In the app today, if you open a folder for a TLF that hasn't yet been
created, then you upload a file into it (thus creating the file), the
view of the folder in the GUI won't be refreshed.  That's because the
initial listing can't subscribe to the TLF before it exists (since
there's no TLF ID yet).

Instead, remember which empty folder couldn't be subscribed to, and
after any operation that may have created the TLF, attempt to
resubcribe again if it was the same TLF that was previous ignored.

Issue: KBFS-4057